Mouth and Face - set size requests
Mouth was missing for the welcome utterance.

Cause was mismanagement of the size after port to GTK 3; it was one
pixel height during the welcome.

Mouth was designed for one on two and a half times or 40% of the
available height.

Instead of recalculating it when the canvas resizes, set a size request
in proportion to the style grid cell size.

Fixes https://github.com/sugarlabs/speak/issues/21

Tested on Ubuntu 19.04 with Sugar 0.116

Signed-off-by: James Cameron <quozl@laptop.org>